+++ extension/src/models/index.ts
@@ -1,3 +1,4 @@
 export { SECRET_KEYS, PROVIDER_BASE_URL_DEFAULTS, PROVIDER_BASE_URL_KEYS } from "./LlmProviderConfig";
 export type { LlmTestResult } from "./LlmTestResult";
 export type { SettingsState } from "./SettingsState";
+export type { ProxyMode, ProxyConfig, ProxyCredentials, ProxyState, ProxyTestResult, ProxyTestInput } from "./ProxyModels";